WIP project refactor
