+++ library/build.gradle
@@ -1,5 +1,7 @@
 plugins {
-    id 'bio.terra.catalog.java-spring-conventions'
+    id 'bio.terra.catalog.java-library-conventions'
+    id 'io.spring.dependency-management'
+    id 'org.springframework.boot'
     id 'de.undercouch.download'
     id 'com.google.cloud.tools.jib'
     id 'com.srcclr.gradle'
@@ -7,20 +9,20 @@ plugins {
 
     id 'com.gorylenko.gradle-git-properties' version '2.3.1'
     id 'org.liquibase.gradle' version '2.1.0'
+    id 'java-library'
 }
-
-apply from: 'generators.gradle'
-apply from: 'publishing.gradle'
 
 dependencies {
     implementation 'bio.terra:terra-common-lib'
-    implementation 'org.springframework.boot:spring-boot-starter-data-jdbc'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.apache.commons:commons-dbcp2'
     implementation 'org.springframework.retry:spring-retry'
+    implementation 'org.springframework:spring-jdbc'
+    implementation 'org.broadinstitute.dsde.workbench:sam-client_2.13:0.1-9867891-SNAP'
     implementation 'javax.ws.rs:javax.ws.rs-api:2.1.1'
     implementation 'org.postgresql:postgresql'
-    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
-    implementation project(':library')
+    implementation 'com.networknt:json-schema-validator:1.0.73'
+    implementation project(':rawls-client')
+    implementation project(':client')
 
     liquibaseRuntime 'org.liquibase:liquibase-core'
     liquibaseRuntime 'info.picocli:picocli:4.6.1'
