+++ library/src/main/java/Library.java
@@ -1,0 +1,4 @@
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+
+@ConfigurationPropertiesScan("bio.terra.catalog")
+public class Library {}
