+++ library/src/main/java/bio/terra/catalog/common/StorageSystem.java
