+++ library/src/main/java/bio/terra/catalog/common/StorageSystemInformation.java
