+++ library/src/main/java/bio/terra/catalog/config/BeanConfig.java
