+++ library/src/main/java/bio/terra/catalog/config/SchemaConfiguration.java
