+++ library/src/main/resources/schema/development/schema.json
