+++ library/src/test/java/bio/terra/catalog/config/BeanConfigTest.java
