+++ service/src/main/resources/application.yml
@@ -21,12 +21,6 @@ env:
 
 # Below here is non-deployment-specific
 
-# When the target is 'local' the write-config.sh script will generate this properties file. It
-# contains the configuration of the BPM test application. We can use that application in our
-# integration testing to make sure the application code paths are working. However, we do not
-# want it to appear in production environments.
-spring.config.import: optional:file:../config/local-properties.yml
-
 logging.pattern.level: '%X{requestId} %5p'
 
 server:
@@ -44,7 +38,7 @@ spring:
           maxAge: 0
           mustRevalidate: true
         useLastModified: false
-      staticLocations: classpath:/api/
+      staticLocations: classpath:/api/,classpath:/schema/development/
 
 catalog:
   ingress:
@@ -75,7 +69,7 @@ catalog:
     basePath: ${env.rawls.basePath}
 
   schema:
-    basePath: classpath:schema/development/schema.json
+    basePath: schema/development/schema.json
 
 terra.common:
   kubernetes:
